Move isomorphisms / anti-corruption functions into a separate module.
